requirements.txt
cryptography
groq
python-dotenv
streamlit
bcrypt
google-generativeai